3 days later of debugging without a debugging tool: Found the reason to the crash, an int converted to a double and then stringified, would result in string length > 2, which in turn would lead to the expression '2 - propertyIdStr.length()' being less than zero.

+++ DeviceAdapters/CoboltOfficial/Property.cpp
@@ -45,7 +45,7 @@ Property::Property( const Stereotype stereotype, const std::string& name ) :
     stereotype_( stereotype ),
     name_( name )
 {
-    const std::string propertyIdStr = std::to_string( (long double) NextPropertyId_++ );
+    const std::string propertyIdStr = std::to_string( NextPropertyId_++ );
     name_ = ( std::string( 2 - propertyIdStr.length(), '0' ) + propertyIdStr ) + "-" + name;
 }
